+++ app/src/main/java/com/example/yandexsummerschool/ui/screens/addTransactionScreen/common/CommonTransactionUi.kt
@@ -1,113 +1,30 @@
-package com.example.yandexsummerschool.ui.screens.addTransactionScreen
+package com.example.yandexsummerschool.ui.screens.addTransactionScreen.common
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavHostController
 import com.example.yandexsummerschool.R
 import com.example.yandexsummerschool.domain.models.ArticleModel
-import com.example.yandexsummerschool.ui.common.components.BottomNavigationBar
 import com.example.yandexsummerschool.ui.common.components.CustomTimePickerDialog
 import com.example.yandexsummerschool.ui.common.components.DatePicker
 import com.example.yandexsummerschool.ui.common.components.ListItem
 import com.example.yandexsummerschool.ui.common.components.ListItemSize
-import com.example.yandexsummerschool.ui.common.components.LoadingIndicator
-import com.example.yandexsummerschool.ui.common.components.TopAppBar
 import com.example.yandexsummerschool.ui.common.components.textFieldColors
-import com.example.yandexsummerschool.ui.common.screens.ErrorScreen
+import com.example.yandexsummerschool.ui.screens.addTransactionScreen.creator.AddTransactionScreenState
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.toImmutableList
-
-@Composable
-fun AddTransactionScreen(
-    viewModelFactory: ViewModelProvider.Factory,
-    navController: NavHostController,
-    isIncome: Boolean,
-) {
-    val viewModel: AddTransactionScreenViewModel = viewModel(factory = viewModelFactory)
-    val state by viewModel.state.collectAsStateWithLifecycle()
-    val accountName by viewModel.accountName.collectAsStateWithLifecycle()
-    var showTimePicker by remember { mutableStateOf(false) }
-    var showDatePicker by remember { mutableStateOf(false) }
-    var isEditingAmount by remember { mutableStateOf(false) }
-    var showArticlesSheet by remember { mutableStateOf(false) }
-    var isEditingComment by remember { mutableStateOf(false) }
-    val articles by viewModel.articles.collectAsStateWithLifecycle()
-
-    LaunchedEffect(Unit) {
-        viewModel.setIncomeType(isIncome)
-        viewModel.loadArticles(isIncome)
-    }
-
-    Scaffold(
-        topBar = {
-            AddTransactionTopBar(
-                isIncome = isIncome,
-                onCancelClick = { navController.popBackStack() },
-                onOkClick = {
-                    viewModel.createTransaction()
-                    navController.popBackStack()
-                },
-            )
-        },
-        bottomBar = { BottomNavigationBar(navController) },
-    ) { paddingValues ->
-        Column(
-            modifier = Modifier.padding(paddingValues),
-        ) {
-            when (val currentState = state) {
-                is AddTransactionScreenState.Error -> ErrorScreen(currentState.message)
-                AddTransactionScreenState.Loading -> LoadingIndicator()
-                is AddTransactionScreenState.Content -> {
-                    AddTransactionScreenContent(
-                        state = currentState,
-                        accountName = accountName,
-                        isEditingAmount = isEditingAmount,
-                        onEditAmount = { isEditingAmount = it },
-                        isEditingComment = isEditingComment,
-                        onEditComment = { isEditingComment = it },
-                        showTimePicker = showTimePicker,
-                        onShowTimePicker = { showTimePicker = it },
-                        showDatePicker = showDatePicker,
-                        onShowDatePicker = { showDatePicker = it },
-                        showArticlesSheet = showArticlesSheet,
-                        onShowArticlesSheet = { showArticlesSheet = it },
-                        onAmountChange = viewModel::changeAmount,
-                        onCommentChange = viewModel::changeComment,
-                        onDateChange = viewModel::changeDate,
-                        onTimeChange = viewModel::changeTime,
-                        onCategoryChange = viewModel::changeCategory,
-                        articles = articles.toImmutableList(),
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 @Suppress("LongParameterList")
-fun AddTransactionScreenContent(
+fun TransactionEditorScreenContent(
     state: AddTransactionScreenState.Content,
     accountName: String,
     isEditingAmount: Boolean,
@@ -212,7 +129,7 @@ fun AmountEditor(amount: String, isEditing: Boolean, onEdit: (Boolean) -> Unit, 
             modifier = Modifier.fillMaxWidth().height(ListItemSize.BIG.size),
             singleLine = true,
             keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done, keyboardType = KeyboardType.Number),
-            keyboardActions = KeyboardActions(onDone = { onEdit(false) }),
+            keyboardActions = androidx.compose.foundation.text.KeyboardActions(onDone = { onEdit(false) }),
             colors = textFieldColors(),
         )
     } else {
@@ -257,7 +174,7 @@ fun CommentEditor(comment: String?, isEditing: Boolean, onEdit: (Boolean) -> Uni
             modifier = Modifier.fillMaxWidth().height(ListItemSize.BIG.size),
             singleLine = false,
             keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done, keyboardType = KeyboardType.Text),
-            keyboardActions = KeyboardActions(onDone = { onEdit(false) }),
+            keyboardActions = androidx.compose.foundation.text.KeyboardActions(onDone = { onEdit(false) }),
             colors = textFieldColors(),
         )
     } else {
@@ -289,15 +206,4 @@ fun ArticlesSheet(
             }
         }
     }
-}
-
-@Composable
-fun AddTransactionTopBar(isIncome: Boolean, onCancelClick: () -> Unit, onOkClick: () -> Unit) {
-    TopAppBar(
-        title = if (isIncome) "Мои доходы" else "Мои расходы",
-        leadingIcon = painterResource(R.drawable.x_icon),
-        onLeadingClick = onCancelClick,
-        trailingIcon = painterResource(R.drawable.ok_icon),
-        onTrailingClick = onOkClick,
-    )
 }
